fix bugs: fix some bugs

+++ internal/typ/resp/UserResp.go
@@ -9,7 +9,7 @@ type LoginResp struct {
 // UserDetailResp 用户个人信息响应
 type UserDetailResp struct {
 	Account string `json:"account"`
-	Role    int64  `json:"role"`
+	Role    int    `json:"role"`
 	Name    string `json:"name"`
 	Avatar  string `json:"avatar"`
 }
